Fixed compile errors on Windows (2).

+++ src/Graphics/Vulkan/Render/GraphicsPipeline.cpp
@@ -117,7 +117,7 @@ void GraphicsPipelineInfo::_resizeColorAttachments(size_t newCount) {
             colorBlendAttachment.colorWriteMask =
                     VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT
                     | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
-            setBlendMode(BlendMode::OVERWRITE, attachmentIdx);
+            setBlendMode(BlendMode::OVERWRITE, uint32_t(attachmentIdx));
         }
         colorBlendInfo.attachmentCount = uint32_t(colorBlendAttachments.size());
         colorBlendInfo.pAttachments = colorBlendAttachments.data();
